change dataTYpe

+++ nasnav-services/src/main/resources/db/changelog/change-logs/00079-20231228-add-usageLimiterCount-coulmn-to-promotionEntity.sql
@@ -5,4 +5,4 @@
 --comment: add usageLimiterCount to promotion table to limit the usage of the promoCode
 
 ALTER TABLE promotions
-    ADD COLUMN IF NOT EXISTS usage_limiter_count BIGINT NULL;
+    ADD COLUMN IF NOT EXISTS usage_limiter_count INTEGER NULL;
